Fix cppcheck warnings, typos, style issues (#1227)
Reducing CppCheck report size, frequent warnings:
* missing override (CI-clang seems also affected by this)
* different argument name
* variable scope can be reduced
* Ogre or containers in standard library most of times use size_t, better to use
* Fix typos
* In general tried to follow the rule from c++ core guidelines https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#c128-virtual-functions-should-specify-exactly-one-of-virtual-override-or-final (but changing it in all code could make PR much bigger)

(cherry picked from commit a395a9527c83dffcdedb376575d80bc00c0c4661)

+++ ogre/src/OgreLidarVisual.cc
@@ -156,7 +156,7 @@ void OgreLidarVisual::SetPoints(const std::vector<double> &_points)
 {
   this->dataPtr->lidarPoints = _points;
   this->dataPtr->pointColors.clear();
-  for (unsigned int i = 0u; i < this->dataPtr->lidarPoints.size(); ++i)
+  for (size_t i = 0u; i < this->dataPtr->lidarPoints.size(); ++i)
   {
     this->dataPtr->pointColors.push_back(gz::math::Color::Blue);
   }
